% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston}
\alias{boston}
\title{Boston housing}
\format{A data frame with 506 rows and 14 columns:
\describe{
\item{\code{crim}}{per capita crime rate by town}
\item{\code{zn}}{proportion of residential land zoned for lots over 25,000 sq.ft}
\item{\code{indus}}{proportion of non-retail business acres per town}
\item{\code{chas}}{Charles River dummy variable (= 1 if tract bounds river; 0 otherwise)}
\item{\code{nox}}{nitrogen oxides concentration (parts per 10 million)}
\item{\code{rm}}{average number of rooms per dwelling}
\item{\code{age}}{proportion of owner-occupied units built prior to 1940}
\item{\code{dis}}{weighted mean of distances to five Boston employment centres}
\item{\code{rad}}{index of accessibility to radial highways}
\item{\code{tax}}{full-value property-tax rate per \$10,000}
\item{\code{ptratio}}{pupil-teacher ratio by town}
\item{\code{black}}{1000(Bk - 0.63)^2 where Bk is the proportion of blacks by town}
\item{\code{lstat}}{lower status of the population (percent)}
\item{\code{medv}}{target variable, median value of owner-occupied homes in \$1000s}
}}
\source{
\code{library(MASS)}

Harrison, D. and Rubinfeld, D.L. (1978) Hedonic prices and the demand for clean air. J. Environ. Economics and Management 5, 81 - 102

Belsley D.A., Kuh, E. and Welsch, R.E. (1980) Regression Diagnostics. Identifying Influential Data and Sources of Collinearity. New York: Wiley.
}
\usage{
data(boston)
}
\description{
A dataset that contains entries that shows median housing values for Boston.
}
\keyword{datasets}
