% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{classifIcePlot}
\alias{classifIcePlot}
\title{ICE Plot}
\usage{
classifIcePlot(pred, var, knots, lines, centered, centerpoint)
}
\arguments{
\item{pred}{The prediction from \code{marginalPrediction()}}

\item{var}{The input variables}

\item{knots}{The amout of knots}

\item{lines}{The amout of lines}

\item{centered}{Selection if ICE plot should be centered or not}

\item{centerpoint}{If ICE plot is centered, the centerpoint can be choosen}
}
\description{
This function generates the Individual Conditional Expectation (ICE) Plot
for classification tasks.
Use \code{imlplots(data, task, models)} as main function and for creating the
interactive Shiny dashboard.
}
