% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{classifPartialDependencePlot}
\alias{classifPartialDependencePlot}
\title{PDP Plot}
\usage{
classifPartialDependencePlot(pred, var, target, knots)
}
\arguments{
\item{pred}{The prediction from \code{marginalPrediction()}}

\item{var}{The input variables}

\item{target}{The target variable}

\item{knots}{The amout of knots}
}
\description{
This function generates the Partial Dependence Plots (PDP)
for classification tasks.
Use \code{imlplots(data, task, models)} as main function and for creating the
interactive Shiny dashboard.
}
