% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imlplots.R
\name{imlplots}
\alias{imlplots}
\title{Create IML plots}
\usage{
imlplots(data, task, models)
}
\arguments{
\item{data}{The data frame used for classification or regression model}

\item{task}{The mlr task, e.g. iris.task = makeClassifTask(data = iris,
target = "Species")}

\item{models}{A list of mlr machine learning models, e.g. list(rf.mod, glm.mod)}
}
\description{
With the imlplots() function an interactive Shiny Dashboard can be created. \cr
Mandatory is the usage of the mlr package, for more infos go to
\url{https://github.com/mlr-org} \cr
Possible are classification or regression tasks.
And three types of plots: Partial dependece plots (PDP), indiviudal conditional
expectation (ICE) plots and accumulated local effects (ALE) plots
}
\note{
You need to install the GitHub version of ggplot2 to run the Shiny App
successfully: devtools::install_github('hadley/ggplot2')
If this version is not installed the plots are not displayed correctly
for less than 30 knots.

Models with the same learner need to have an unique ID
Unique ID's can be given via the makeLearner function
E.g. makeLearner("regr.randomForest", id = "ownId")
}
\author{
Julia Fried, Tobias Riebe, Christian Scholbeck
}
