% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{makePredictionsSampled}
\alias{makePredictionsSampled}
\title{Generate Prediction Sample}
\usage{
makePredictionsSampled(data, var, knots, lines, model, type)
}
\arguments{
\item{data}{The input data for model}

\item{var}{The variables of model}

\item{knots}{The amout of knots}

\item{lines}{The amout of lines}

\item{model}{The fitted model}

\item{type}{The type of learning problem (classification or regression)}
}
\description{
This function generates samples for the prediction function.
}
\note{
The main function is \code{imlplots(data, task, models)}. Get more
information with \code{?imlplots}.
}
