% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modelCheck}
\alias{modelCheck}
\title{Model check function}
\usage{
modelCheck(data, var, models)
}
\arguments{
\item{data}{The input data for `marginalPrediction()`}

\item{var}{The variables of data}

\item{models}{One or several models that are checked from this function}
}
\description{
Some mlr learners, like \code{classif.lda} or \code{classif.qda}, can't be
used in the marginalPrediction function from the \pkg{mmpf} package.
This \code{modelCheck()} allows only the usage of models that work together
with the \pkg{mmpf} package.
}
\note{
The main function is \code{imlplots(data, task, models)}, check out
\code{?imlplots} for more information.
}
