% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{regrAlePlot}
\alias{regrAlePlot}
\title{ALE Plot}
\usage{
regrAlePlot(data, model, target, var, knots)
}
\arguments{
\item{target}{The target variable}

\item{var}{The input variables}

\item{knots}{The amout of knots}

\item{pred}{The prediction from \code{marginalPrediction()}}
}
\description{
Create an accumulated local effects plot for regression tasks.
Use \code{imlplots(data, task, models)} as main function and for creating the
interactive Shiny dashboard.
}
